requirements/base.txt
channels
django
django-environ
django-webpack-loader
djangorestframework
psycopg2
